Add ParserHelper files to project
Move error reporting methods to ParserHelper.
