+++ NAS2D/Configuration.cpp
@@ -10,6 +10,7 @@
 
 #include "Configuration.h"
 #include "ContainerUtils.h"
+#include "ParserHelper.h"
 #include "Filesystem.h"
 #include "Utility.h"
 #include "Xml/Xml.h"
@@ -22,25 +23,6 @@ using namespace NAS2D;
 
 
 namespace {
-	void reportMissingOrUnexpected(const std::vector<std::string>& missing, const std::vector<std::string>& unexpected)
-	{
-		if (!missing.empty() || !unexpected.empty())
-		{
-			const auto missingString = !missing.empty() ? "Missing names: {" + join(missing, ", ") +"}" : "";
-			const auto unexpectedString = !unexpected.empty() ? "Unexpected names: {" + join(unexpected, ", ") + "}" : "";
-			const auto joinString = (!missingString.empty() && !unexpectedString.empty()) ? "\n" : "";
-			throw std::runtime_error(missingString + joinString + unexpectedString);
-		}
-	}
-
-	void reportMissingOrUnexpected(const std::vector<std::string>& names, const std::vector<std::string>& required, const std::vector<std::string>& optional)
-	{
-		const auto missing = missingValues(names, required);
-		const auto unexpected = unexpectedValues(names, required, optional);
-		reportMissingOrUnexpected(missing, unexpected);
-	}
-
-
 	Dictionary attributesToDictionary(const Xml::XmlElement& element)
 	{
 		Dictionary dictionary;
